requirements.txt
# Not all of these are relevant to this particular implementation

streamlit
openai==1.3.0
pdfminer.six
nltk
tqdm
gpt-index
streamlit-chat
stop-words
Pillow
scikit-learn
streamlit-extras
requests
pinecone-client==3.0.0
pinecone-datasets
pinecone-text
unstructured==0.10.24
sentence-transformers==2.2.2
llama-index==0.9.34
pdf2image==1.16.3
python-dotenv==1.0.0
pytesseract==0.3.10
unstructured_pytesseract==0.3.12